requirements.txt
# Core API framework
fastapi>=0.115.0

# ASGI server
uvicorn>=0.30.0

# Data validation (FastAPI compatible)
pydantic>=2.9.0

# Redis client (sync, stable)
redis>=5.0.0

# Optional but recommended (internal FastAPI dependency safety)
typing-extensions>=4.9.0

# Agentic Orchestration
crewai>=0.1.0
replit-ai>=0.1.0
python-dotenv>=1.0.0
prometheus-client>=0.17.0
